requirements.txt
openai
langchain
langchain-openai
langchain-chroma
python-dotenv
pandas